Impelement the GameManager.cs.
Create Hit animation for Player.
Create DeathZone.cs for Impelement the logic of Die and Lose Player when fall of the ground.
Create Final.cs for Impelement the logic of Win when the player arrive and collect the End item in scene.
Create GameManager.cs for manage the player win and lose .
Create Spawner.cs for Spawn the player when the room is created.
Build the game for test.
Add red-heart.png for use the Health Ui.(Will be used later)